requirements.txt
# AI and API dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: For better CLI experience
click>=8.0.0
rich>=13.0.0

